+++ tests/aws/test-class_aws_qs.R
@@ -1,4 +1,4 @@
-# Use sparingly. We do not want to max out any AWS quotas.
+# Use sparingly to minimize AWS costs.
 # And afterwards, manually verify that all the buckets are gone.
 tar_test("aws_qs format data gets stored", {
   skip_if_no_aws()