+++ tests/testthat/test-utils_data.R
@@ -22,3 +22,35 @@ tar_test("keyvalue_field()", {
     "sdfasdf"
   )
 })
+
+tar_test("omit_null()", {
+  x <- list(e = 1L, a = NULL, c = 2L, d = NULL, b = 3L)
+  expect_equal(length(x), 5)
+  expect_equal(names(x), c("e", "a", "c", "d", "b"))
+  y <- omit_null(x)
+  expect_equal(y, list(e = 1L, c = 2L, b = 3L))
+  expect_equal(length(y), 3)
+  expect_equal(names(y), c("e", "c", "b"))
+})
+
+tar_test("supported_args()", {
+  f <- function(a, b, c = "c_default", d = "d_default", x = "x_default") {
+  }
+  args <- list(
+    a = "a_value",
+    b = "b_value",
+    c = "c_value",
+    d = NULL,
+    e = "nope",
+    f = NULL
+  )
+  out <- supported_args(fun = f, args = args)
+  expect_equal(
+    out,
+    list(
+      a = "a_value",
+      b = "b_value",
+      c = "c_value"
+    )
+  )
+})